tests/wpt/tests/css/css-text/parsing/word-boundary-detection-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().wordBoundaryDetection</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundary-detection">
<meta name="assert" content="word-boundary-detection computed value is specified keyword.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("word-boundary-detection", "normal");
test_computed_value("word-boundary-detection", "manual");
/* Not sure how to test for auto(<lang>) generically, as which one will parse depends on languages actually supported */
</script>
</body>
</html>
